requirements.txt
pyvista==0.44.2
python-dotenv==1.0.1
numpy>=1.24.0
vtk==9.4.2
